crs/src/crs-cp-java/requirements.txt
pyyaml
coloredlogs
aiofiles